api/requirements.txt
-e .
boto3==1.34.131
credit-score-mlops==0.0.1
dacite==1.6.0
dataclasses-json==0.6.7
email-validator==2.2.0
ensure==1.0.4
fastapi==0.112.0
fastapi-cli==0.0.5
flufl-lock==8.1.0
fusepy==3.0.1
gql==3.5.0
httptools==0.6.1
httpx==0.23.3
ipykernel==6.29.5
loguru==0.7.2
markdown-it-py==3.0.0
mlflow==2.15.0
pathvalidate==3.0.0
pip-chill==1.0.3
python-box==7.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
requests-toolbelt==1.0.0
treelib==1.6.4
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0